# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: "A.R blog api"
  description: |
    The A.R Blog API is a robust and flexible RESTful API that allows you to manage blog posts, comments, and user accounts for your blogging platform. It provides a secure and efficient way to create, retrieve, update, and delete blog content.

    **Key Features:**
    - Create and manage blog posts with titles, content, and tags.
    - Allow user registration and authentication for creating and managing their own blog posts and comments.
    - Enable commenting on blog posts to foster community engagement.
    - Retrieve lists of blog posts with pagination and filtering options.
    - Built-in security measures, including authentication and authorization, to protect user data and content
  contact:
    name: "A.R Tech Hub"
    url: "https//a.r.i.tech-venture"
    email: arifulislam10111999@gmail.com
servers:
  - description: dev
    url: http://localhost:5000/api/v1
  - description: server
    url: https://ar-blog-api.onrender.com/api/v1
...
paths:
...
  "/auth/signin":
    post:
      description: sing in by credadintail
      tags:
        - Auth
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: ariful@example.com
                password:
                  type: string
                  format: password
                  example: pass123
              required: [email, password]
      responses:
        "200":
          description: sing in successfuly
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: sign in successful
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        example: "yJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"

                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        example: "/auth/singup"

        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"

    get:
      summary: Retrieve Signin Information
      tags:
        - Auth
      description: Retrieve information about the signin process.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A success message.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message indicating that the resource was not found.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message indicating an internal server error.
                description: sing in by credadintail
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    # bearerAuth:
    #   type: http
    #   scheme: bearer
    #   bearerFormat: JWT
...
  responses:
    "500":
      description: Internal Server Error
      content:
        "aplication/json":
          schema:
            type: object
            properties:
              code:
                type: string
                format: int32
                example: 500
              error:
                type: string
                example: Internal Server Error
              message:
                type: string
                example: we are sory for the inconvenent . pleace try agin later
    "400":
      description: bad request
      content:
        "aplication/json":
          schema:
            type: object
            properties:
              code:
                type: string
                format: int32
                example: 400
              error:
                type: string
                example: bad request
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: email
                    message:
                      type: string
                      example: invalide email
                    in:
                      type: string
                      enum: [query.body]
                      example: body
...